Fixed incorrect health for delay option and health display values
Fixed:
- Delay option not getting the right health
- Showing negative health number when the entity dies
- Showing decimals

+++ src/com/zeshanaslam/actionhealth/HealthListeners.java
@@ -63,7 +63,7 @@ public class HealthListeners implements Listener {
 
                 // Send health
                 LivingEntity livingEntity = (LivingEntity) damaged;
-                plugin.healthUtil.sendHealth(player, (LivingEntity) damaged, livingEntity.getHealth() - event.getFinalDamage());
+                plugin.healthUtil.sendHealth(player, (LivingEntity) damaged, (int) (livingEntity.getHealth() - event.getFinalDamage()));
 
             }
         }
@@ -92,7 +92,7 @@ public class HealthListeners implements Listener {
 
             // Send health
             LivingEntity livingEntity = (LivingEntity) damaged;
-            plugin.healthUtil.sendHealth(player, (LivingEntity) damaged, livingEntity.getHealth() - event.getFinalDamage());
+            plugin.healthUtil.sendHealth(player, (LivingEntity) damaged, (int) (livingEntity.getHealth() - event.getFinalDamage()));
         }
     }
